move import models to own folder
