+++ services/import/fantasyTeamService.js
@@ -2,7 +2,7 @@ import pool from '../db.js';
 import yahooApiService from './yahooApiService.js';
 import ownerService from './ownerService.js';
 import leagueService from './leagueService.js';
-import fantasyTeamModel from '../../models/fantasyteamModel.js';
+import fantasyTeamModel from '../../models/import/fantasyteamModel.js';
 
 const seasonsToImport = async () => {
   try {
