+++ services/import/gameCodeTypeService.js
@@ -1,6 +1,6 @@
 import pool from '../db.js';
 import yahooService from './yahooApiService.js';
-import gameCodeTypeModel from '../../models/gamecodeTypeModel.js';
+import gameCodeTypeModel from '../../models/import/gamecodeTypeModel.js';
 
 async function getAllCodeTypes() {
   try {
